requirements.txt
pandas 1.4.2
matplotlib.pyplot  3.7.1
seaborn 0.11.2
geopandas 0.12.2
folium 0.14.0 
pycirclize 0.3.0
sklearn.model_selection. train_test_split    
sklearn.metrics.classification_report
sklearn.preprocessing.OrdinalEncoder
xgboost.XGBClassifier  1.7.4
streamlit  1.20.0
streamlit-folium 0.11.1

funciones.kwichon - librería de funciones para este proyecto

Descargar carpeta 'datos' que contiene archivos con las migraciones de: https://drive.google.com/drive/folders/1QGW4sFkmz6wWSkeby4oPHDzKO6STyRiB?usp=sharing
